backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Queue
redis==5.0.1
rq==1.15.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-social-auth==0.3.6
google-auth==2.26.2
google-auth-oauthlib==1.2.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# AWS/Storage
boto3==1.34.23
botocore==1.34.23

# Rate Limiting
slowapi==0.1.9

# Monitoring and Logging
sentry-sdk[fastapi]==1.39.2
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pillow==10.2.0
requests==2.31.0
